+++ lib/src/noise/white.dart
@@ -5,14 +5,12 @@ class WhiteNoise {
 
   WhiteNoise({this.seed: 1337});
 
-  double getWhiteNoise2(double x, double y) =>
-      valCoord2D(seed, x.toInt(), y.toInt());
+  double getWhiteNoise2(int x, int y) => valCoord2D(seed, x, y);
 
-  double getWhiteNoise3(double x, double y, double z) =>
-      valCoord3D(seed, x.toInt(), y.toInt(), z.toInt());
+  double getWhiteNoise3(int x, int y, int z) => valCoord3D(seed, x, y, z);
 
-  double getWhiteNoise4(double x, double y, double z, double w) =>
-      valCoord4D(seed, x.toInt(), y.toInt(), z.toInt(), w.toInt());
+  double getWhiteNoise4(int x, int y, int z, int w) =>
+      valCoord4D(seed, x, y, z, w);
 
   double getWhiteNoiseInt2(int x, int y) {
     return valCoord2D(seed, x, y);
